chore: experimental conformance promotion
